Stop using __LINE_BOT_SDK_PYTHON_VERSION__ (#710)
The string `__LINE_BOT_SDK_PYTHON_VERSION__ ` is used in
`linebot/__about__.py` to make it easier to replace with the new version
at the time of release. However, this string is inappropriate as a
version, so it needs to be replaced when testing locally... This is
cumbersome for us, so this PR addresses that issue.

At the time of release, the file that need to be updated will be
directly modified anyway. `./tools/update_version.py` will handle this.
Since 1 lines across 1 files need to be changed, the CI will fail if
this condition is not met. Reviewers can verify that the version is
changed locally, but it's up to them whether to do so.