Limit the timing of git submodule updates by renovate to avoid disrupting work during working hours (#743)
After updating line-openapi, there are times when both
- (1) a PR generated after Renovate updates the git submodule and
- (2) a PR created by an operator dispatching the GitHub workflow

are generated simultaneously.
This change is to avoid that. Usually, when it's known that the bot SDK
code will be changed (which is most cases), (1) is not necessary because
the operator dispatches the workflow to change the code.

In particular, since the operator usually dispatches the workflow for 6
bot sdk repositories via a script, the PRs automatically created by
Renovate are a hindrance.

However, to account for the possibility that the operator might simply
forget to update bot sdk code, I've restricted Renovate to update the
git submodule only at night instead of not updating. This way, if code
generation is forgotten, it can be noticed the next day.

(Other options) It was also possible to change the GitHub workflow to
prevent duplicate PR by using same branch, but if the operator forgets
to update for a while, there will be multiple features changed in one
PR. I don't like this. though both options are fine, I think this PR is
more suitable for our work at the moment rather than this option.